rule-install-rate: add support for multiple CPUs
As we are seeing how this scales after the rtnl_lock removal, this now
supports dealing with multiple CPUs. This is not triggered by the
iproute test but with another tool that will soon be added.

In order to better see how the time is spent, it is summing the time
used by all CPUs. With this, if it faces a contention, it will have
larger drifts on x axis.