rule-install-rate: run tc always on the same CPU
The 'tc flower cumulative' line tracks also the time spent between fl_change
calls. When running without multi-CPU support, as tc is single threaded, it
made no difference on which CPU it was running.

But now that we support multi-CPU, it is also considering the idle time on CPUs
that it once executed and migrated back, and the curve gets "steppy".

To avoid it, as tc is single threaded anyway, pin it to a single CPU.