perf-plot: use a dynamic y-axis maximum for rate graph
There can be too much noise in the beginning of the sample, causing it
to flatten the curve. If we make the ceiling too low, it won't cover
some tests. Too high, as explained. So lets use a dynamic value based on
a measured average and give it some room (4 times) so we can see spikes.