perf-plot.sh: fix issue with unbound entries
Currently, the grep only matches on the entry and exit points, without
worrying if there is an exit point for a given entry. For stats graph
that's a problem because it may get cut in the middle, and then the .dat
file gets corrupted.

Fix it by using a sed expression that only prints the entry point IF
there is a corresponding exit one. As it was already sorted by process
id, that's enough.